+++ CHANGELOG.md
@@ -6,6 +6,15 @@ SPDX-License-Identifier: MIT
 
 # Changelog
 
+## tol-sdk `2.5.2`
+
+10-09-2026
+
+- Added `SetRelationshipAction` to set (or clear, via a null `related_id`) a to-one relationship on one or more objects
+- Corrected issues with enriched provenanced fields
+- Corrected bug with [] field selectors
+- Make specimen a proper relationship from tissue_prep
+
 ## tol-sdk `2.5.1`
 
 02-09-2026